Split modules tests to safe, unsafe, and utils
