+++ tests/modules.safe.test.js
@@ -746,457 +746,3 @@ describe('SAFE: simplifyIfStatements', async () => {
 		assert.strictEqual(result, expected);
 	});
 });
-
-describe('UNSAFE: normalizeRedundantNotOperator', async () => {
-	const targetModule = (await import('../src/modules/unsafe/normalizeRedundantNotOperator.js')).default;
-	it('TP-1', () => {
-		const code = `!true || !false || !0 || !1 || !a || !'a' || ![] || !{} || !-1 || !!true || !!!true`;
-		const expected = `false || true || true || false || !a || false || false || false || false || true || false;`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-describe('UNSAFE: resolveAugmentedFunctionWrappedArrayReplacements', async () => {
-	// Load the module even though there are no tests for it - to include it in the coverage report
-	// noinspection JSUnusedLocalSymbols
-	const targetModule = (await import('../src/modules/unsafe/resolveAugmentedFunctionWrappedArrayReplacements.js')).evalWithDom;
-	it.todo('TODO: Write tests for function', () => {});
-});
-describe('UNSAFE: resolveBuiltinCalls', async () => {
-	const targetModule = (await import('../src/modules/unsafe/resolveBuiltinCalls.js')).default;
-	it('TP-1: atob', () => {
-		const code = `atob('c29sdmVkIQ==');`;
-		const expected = `'solved!';`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TP-2: btoa', () => {
-		const code = `btoa('solved!');`;
-		const expected = `'c29sdmVkIQ==';`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TP-3: split', () => {
-		const code = `'ok'.split('');`;
-		const expected = `[\n  'o',\n  'k'\n];`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TN-1: querySelector', () => {
-		const code = `document.querySelector('div');`;
-		const expected = code;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TN-2: Unknown variable', () => {
-		const code = `atob(x)`;
-		const expected = code;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TN-3: Overwritten builtin', () => {
-		const code = `function atob() {return 1;} atob('test');`;
-		const expected = code;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-describe('UNSAFE: resolveDefiniteBinaryExpressions', async () => {
-	const targetModule = (await import('../src/modules/unsafe/resolveDefiniteBinaryExpressions.js')).default;
-	it('TP-1', () => {
-		const code = `5 * 3; '2' + 2; '10' - 1; 'o' + 'k'; 'o' - 'k'; 3 - -1;`;
-		const expected = `15;\n'22';\n9;\n'ok';\nNaN;\n4;`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-describe('UNSAFE: resolveDefiniteMemberExpressions', async () => {
-	const targetModule = (await import('../src/modules/unsafe/resolveDefiniteMemberExpressions.js')).default;
-	it('TP-1', () => {
-		const code = `'123'[0]; 'hello'.length;`;
-		const expected = `'1';\n5;`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TN-1', () => {
-		const code = `++[[]][0];`;
-		const expected = code;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-describe('UNSAFE: resolveDeterministicConditionalExpressions', async () => {
-	const targetModule = (await import('../src/modules/unsafe/resolveDeterministicConditionalExpressions.js')).default;
-	it('TP-1', () => {
-		const code = `(true ? 1 : 2); (false ? 3 : 4);`;
-		const expected = `1;\n4;`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TN-1', () => {
-		const code = `({} ? 1 : 2); ([].length ? 3 : 4);`;
-		const expected = code;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-describe('UNSAFE: resolveEvalCallsOnNonLiterals', async () => {
-	const targetModule = (await import('../src/modules/unsafe/resolveEvalCallsOnNonLiterals.js')).default;
-	it('TP-1', () => {
-		const code = `eval(function(a) {return a}('atob'));`;
-		const expected = `atob;`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TP-2', () => {
-		const code = `eval([''][0]);`;
-		const expected = `''`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-describe('UNSAFE: resolveFunctionToArray', async () => {
-	const targetModule = (await import('../src/modules/unsafe/resolveFunctionToArray.js')).default;
-	it('TP-1', () => {
-		const code = `function a() {return [1];}\nconst b = a();`;
-		const expected = `function a() {\n  return [1];\n}\nconst b = [1];`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-describe('UNSAFE: resolveInjectedPrototypeMethodCalls', async () => {
-	const targetModule = (await import('../src/modules/unsafe/resolveInjectedPrototypeMethodCalls.js')).default;
-	it('TP-1', () => {
-		const code = `String.prototype.secret = function () {return 'secret ' + this;}; 'hello'.secret();`;
-		const expected = `String.prototype.secret = function () {\n  return 'secret ' + this;\n};\n'secret hello';`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-describe('UNSAFE: resolveLocalCalls', async () => {
-	const targetModule = (await import('../src/modules/unsafe/resolveLocalCalls.js')).default;
-	it('TP-1: Function declaration', () => {
-		const code = `function add(a, b) {return a + b;} add(1, 2);`;
-		const expected = `function add(a, b) {\n  return a + b;\n}\n3;`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TP-2: Arrow function', () => {
-		const code = `const add = (a, b) => a + b; add(1, 2);`;
-		const expected = `const add = (a, b) => a + b;\n3;`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TP-3: Overwritten builtin', () => {
-		const code = `const atob = (a, b) => a + b; atob('got-');`;
-		const expected = `const atob = (a, b) => a + b;\n'got-undefined';`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TN-1: Missing declaration', () => {
-		const code = `add(1, 2);`;
-		const expected = code;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TN-2: Skipped builtin', () => {
-		const code = `btoa('a');`;
-		const expected = code;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TN-2: No replacement with undefined', () => {
-		const code = `function a() {} a();`;
-		const expected = code;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-describe('UNSAFE: resolveMinimalAlphabet', async () => {
-	const targetModule = (await import('../src/modules/unsafe/resolveMinimalAlphabet.js')).default;
-	it('TP-1', () => {
-		const code = `+true; -true; +false; -false; +[]; ~true; ~false; ~[]; +[3]; +['']; -[4]; ![]; +[[]];`;
-		const expected = `1;\n-'1';\n0;\n-0;\n0;\n-'2';\n-'1';\n-'1';\n3;\n0;\n-'4';\nfalse;\n0;`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TP-2', () => {
-		const code = `[] + []; [+[]]; (![]+[]); +[!+[]+!+[]];`;
-		const expected = `'';\n[0];\n'false';\n2;`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TN-1', () => {
-		const code = `-false; -[]; +{}; -{}; -'a'; ~{}; -['']; +[1, 2]; +this; +[this];`;
-		const expected = `-0;\n-0;\n+{};\n-{};\nNaN;\n~{};\n-0;\nNaN;\n+this;\n+[this];`;
-		const result = applyModuleToCode(code, targetModule);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-
-describe('UTILS: evalInVm', async () => {
-	const targetModule = (await import('../src/modules/utils/evalInVm.js')).evalInVm;
-	it('TP-1', () => {
-		const code = `'hello ' + 'there';`;
-		const expected = {type: 'Literal', value: 'hello there', raw: 'hello there'};
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TN-1', () => {
-		const code = `Math.random();`;
-		const expected = badValue;
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TN-2', () => {
-		const code = `function a() {return console;} a();`;
-		const expected = badValue;
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-describe('UTILS: evalWithDom', async () => {
-	// Load the module even though there are no tests for it - to include it in the coverage report
-	// noinspection JSUnusedLocalSymbols
-	const targetModule = (await import('../src/modules/utils/evalWithDom.js')).evalWithDom;
-	it.todo('TODO: Write tests for function', () => {});
-});
-describe('UTILS: areReferencesModified', async () => {
-	const targetModule = (await import('../src/modules/utils/areReferencesModified.js')).areReferencesModified;
-	it('TP-1', () => {
-		const code = `let a = 1; let b = 2 + a, c = a + 3; a++;`;
-		const expected = true;
-		const ast = generateFlatAST(code);
-		const result = targetModule(ast, ast.find(n => n.src === 'a = 1').id.references);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TP-2', () => {
-		const code = `let a = 1; let b = 2 + a, c = (a += 2) + 3;`;
-		const expected = true;
-		const ast = generateFlatAST(code);
-		const result = targetModule(ast, ast.find(n => n.src === 'a = 1').id.references);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TN-1', () => {
-		const code = `const a = 1; let b = 2 + a, c = a + 3;`;
-		const expected = false;
-		const ast = generateFlatAST(code);
-		const result = targetModule(ast, ast.find(n => n.src === 'a = 1').id.references);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-describe('UTILS: createNewNode', async () => {
-	const targetModule = (await import('../src/modules/utils/createNewNode.js')).createNewNode;
-	it('Literan: String', () => {
-		const code = 'Baryo';
-		const expected = {type: 'Literal', value: 'Baryo', raw: 'Baryo'};
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('Literal: Number - positive number', () => {
-		const code = 3;
-		const expected = {type: 'Literal', value: 3, raw: '3'};
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('Literal: Number - negative number', () => {
-		const code = -3;
-		const expected =  {type: 'UnaryExpression', operator: '-', argument: {type: 'Literal', value: '3', raw: '3'}};
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('Literal: Number - negative infinity', () => {
-		const code = -Infinity;
-		const expected =  {type: 'UnaryExpression', operator: '-', argument: {type: 'Identifier', name: 'Infinity'}};
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('Literal: Number - NOT operator', () => {
-		const code = '!3';
-		const expected =  {type: 'UnaryExpression', operator: '!', argument: {type: 'Literal', value: '3', raw: '3'}};
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('Literal: Number - Identifier', () => {
-		const code1 = Infinity;
-		const expected1 =  {type: 'Identifier', name: 'Infinity'};
-		const result1 = targetModule(code1);
-		assert.deepStrictEqual(result1, expected1);
-		const code2 = NaN;
-		const expected2 =  {type: 'Identifier', name: 'NaN'};
-		const result2 = targetModule(code2);
-		assert.deepStrictEqual(result2, expected2);
-	});
-	it('Literal: Boolean', () => {
-		const code = true;
-		const expected = {type: 'Literal', value: true, 'raw': 'true'};
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('Array: empty', () => {
-		const code = [];
-		const expected = {type: 'ArrayExpression', elements: []};
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('Array: populated', () => {
-		const code = [1, 'a'];
-		const expected = {type: 'ArrayExpression', elements: [
-			{type: 'Literal', value: 1, raw: '1'},
-			{type: 'Literal', value: 'a', raw: 'a'}
-		]};
-		const result = targetModule(code);
-		assert.deepEqual(result, expected);
-	});
-	it('Object: empty', () => {
-		const code = {};
-		const expected = {type: 'ObjectExpression', properties: []};
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('Object: populated', () => {
-		const code = {a: 1};
-		const expected = {type: 'ObjectExpression', properties: [{
-			type: 'Property',
-			key: {type: 'Literal', value: 'a', raw: 'a'},
-			value: {type: 'Literal', value: 1, raw: '1'}
-		}]};
-		const result = targetModule(code);
-		assert.deepEqual(result, expected);
-	});
-	it('Object: populated with BadValue', () => {
-		const code = {a() {}};
-		const expected = badValue;
-		const result = targetModule(code);
-		assert.deepEqual(result, expected);
-	});
-	it('Undefined', () => {
-		const code = undefined;
-		const expected = {type: 'Identifier', name: 'undefined'};
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-	it('Null', () => {
-		const code = null;
-		const expected = {type: 'Literal', raw: 'null'};
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-	it.todo('TODO: Implement Function', () => {
-	});
-	it('RegExp', () => {
-		const code = /regexp/gi;
-		const expected = {type: 'Literal', regex: {flags: 'gi', pattern: 'regexp'}};
-		const result = targetModule(code);
-		assert.deepStrictEqual(result, expected);
-	});
-
-});
-describe('UTILS: createOrderedSrc', async () => {
-	const targetModule = (await import('../src/modules/utils/createOrderedSrc.js')).createOrderedSrc;
-	it('TP-1: Re-order nodes', () => {
-		const code = 'a; b;';
-		const expected = `a\nb\n`;
-		const ast = generateFlatAST(code);
-		const targetNodes = [
-			4, // b()
-			2, // a()
-		];
-		const result = targetModule(targetNodes.map(n => ast[n]));
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TP-2: Wrap calls in expressions', () => {
-		const code = 'a();';
-		const expected = `a();\n`;
-		const ast = generateFlatAST(code);const targetNodes = [
-			2, // a()
-		];
-		const result = targetModule(targetNodes.map(n => ast[n]));
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TP-3: Push IIFEs to the end in order', () => {
-		const code = '(function(a){})(); a(); (function(b){})(); b();';
-		const expected = `a();\nb();\n(function(a){})();\n(function(b){})();\n`;
-		const ast = generateFlatAST(code);
-		const targetNodes = [
-			10, // (function(b){})()
-			15, // b()
-			7, // a()
-			2, // (function(a){})()
-		];
-		const result = targetModule(targetNodes.map(n => ast[n]));
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TP-4: Add dynamic name to IIFEs', () => {
-		const code = '!function(a){}(); a();';
-		const expected = `a();\n(function func3(a){}());\n`;
-		const ast = generateFlatAST(code);const targetNodes = [
-			3, // function(a){}()
-			8, // a()
-		];
-		const result = targetModule(targetNodes.map(n => ast[n]));
-		assert.deepStrictEqual(result, expected);
-	});
-	it('TP-5: Add variable name to IIFEs', () => {
-		const code = 'const b = function(a){}(); a();';
-		const expected = `a();\n(function b(a){}());\n`;
-		const ast = generateFlatAST(code);const targetNodes = [
-			4, // function(a){}()
-			9, // a()
-		];
-		const result = targetModule(targetNodes.map(n => ast[n]));
-		assert.deepStrictEqual(result, expected);
-	});
-	it(`TP-6: Preserve node order`, () => {
-		const code = '(function(a){})(); a(); (function(b){})(); b();';
-		const expected = `(function(a){})();\na();\n(function(b){})();\nb();\n`;
-		const ast = generateFlatAST(code);
-		const targetNodes = [
-			10, // (function(b){})()
-			7, // a()
-			15, // b()
-			2, // (function(a){})()
-		];
-		const result = targetModule(targetNodes.map(n => ast[n]), true);
-		assert.deepStrictEqual(result, expected);
-	});
-	it(`TP-7: Standalone FEs`, () => {
-		const code = '~function(iife1){}();~function(iife2){}();';
-		const expected = `(function func4(iife1){});\n(function func10(iife2){});\n`;
-		const ast = generateFlatAST(code);
-		const targetNodes = [
-			10, // function(iife2){}
-			4, // function(iife1){}
-		];
-		const result = targetModule(targetNodes.map(n => ast[n]), true);
-		assert.deepStrictEqual(result, expected);
-	});
-});
-describe('UTILS: getCache', async () => {
-	const getCache = (await import('../src/modules/utils/getCache.js')).getCache;
-	it('TP-1: Retain values', () => {
-		const key1 = 'hash1';
-		const key2 = 'hash2';
-		const cache = getCache(key1);
-		assert.deepStrictEqual(cache, {});
-		cache['key1'] = 'value1';
-		const expectedC1 = {key1: 'value1'};
-		assert.deepStrictEqual(cache, expectedC1);
-		const cache2 = getCache(key1);
-		assert.deepStrictEqual(cache2, expectedC1);
-		const cache3 = getCache(key2);
-		assert.deepStrictEqual(cache3, {});
-	});
-	it('TP-2: Flush cache', () => {
-		const key = 'flush1';
-		let cache = getCache(key);
-		assert.deepStrictEqual(cache, {});
-		cache['k'] = 'v';
-		const expectedC1 = {k: 'v'};
-		assert.deepStrictEqual(cache, expectedC1);
-		getCache.flush();
-		cache = getCache(key);
-		assert.deepStrictEqual(cache, {});
-	});
-});
-
